# specs/swaggerdoc.yaml
openapi: 3.0.0
info:
  version: '1.0.0'
  title: 'Medten Product API'
  description: 'Interact with the Medten product database using this API.'
servers:
  - url: https://mgtwitxxi1.execute-api.ap-southeast-1.amazonaws.com/Prod/api/v1/
    description: API Test Environment
  - url: https://v73kddyeqk.execute-api.ap-southeast-1.amazonaws.com/Prod/api/v1/
    description: API POC (Proof of Concept) Environment
  - url: http://localhost:3000/api/v1/
    description: Developer Local Test Environments
  # Added by API Auto Mocking Plugin
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/gaumeister/medten-product-api/1.0.0
paths: 
...
  /products/${id}/families:
    summary: The families for a product.
    parameters:
    - name: id
      in: path
      description: the product id
      required: true
      schema:
        type: integer
    get:
      summary: gets all families for a product
      tags:
        - products
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  product_family_connects:
                    type: array
                    items:  
                      $ref: '#/components/schemas/ProductFamily'
        400:
          $ref: '#/components/responses/400ClientError'
        500:
          $ref: '#/components/responses/500ServerError'
    post:
      summary: replaces the current families for a product
      description: Replaces the current familiest for a product, adding those that need to be added, updating any existing, and removing any that no longer match the payload. Family references are differentiated by `family_id` property.
      tags:
        - products
      requestBody:
        description: >
          The new families for the product
        required: true
        content:
          application/json:
            schema:
              type: array
              items:  
                $ref: '#/components/schemas/ProductFamily'
                  
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:  
                  $ref: '#/components/schemas/ProductFamily'
              
        400:
          $ref: '#/components/responses/400ClientError'
        500:
          $ref: '#/components/responses/500ServerError'
...
  /families:
    summary: Represents a product family
    post:
      summary: creates a family
      tags:
        - families
      requestBody:
        description:  The family to be created.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Family'
      responses:
        200:
          description: returns the family 
          content:
            application/json:
              schema: 
                $ref: '#/components/schemas/Family'
        400:
          $ref: '#/components/responses/400ClientError'
        500:
          $ref: '#/components/responses/500ServerError'
  /families/${id}:
    summary: Represents a product family.
    parameters:
    - name: id
      in: path
      description: the family id
      required: true
      schema:
        type: integer
    get:
      summary: gets family by id
      tags:
        - families
      responses:
        200:
          description: OK
          content:
            application/json:
              schema: 
                $ref: '#/components/schemas/FamilyView'
        400:
          $ref: '#/components/responses/400ClientError'
        500:
          $ref: '#/components/responses/500ServerError'
    put:
      summary: updates a family
      tags:
        - families
      requestBody:
        description: >
          The family to be updated.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Family'
      responses:
        200:
          description: returns the family 
          content:
            application/json:
              schema: 
                $ref: '#/components/schemas/Family'
        400:
          $ref: '#/components/responses/400ClientError'
        500:
          $ref: '#/components/responses/500ServerError'
  /families/search:
    post:
      summary: Perform an advanced search for families.
      description: Extends basic search capabilities, allowing a search term to be used across different family fields. Additional filter criteria can be used to further narrow the search-term search.
      tags:
        - families
      parameters:
        - in: header
          name: Accept
          required: false
          description: If omitted, `application/json` is assumed.
          schema:
            type: string
            enum:
              - application/json
              - text/csv
      requestBody:
        description: >
          The search criteria to use for the search.
        required: true
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/SearchCriteria'
              - type: object
                properties:
                  search_term:
                    type: string
                    description: wildcard search value to be used for searching any/all of the `search_term_fields`
                  search_term_fields:
                    description: which family fields the search term is applied to
                    type: array
                    items:
                      type: string
                      enum: [family_code, oem_brand_en, name_en, category_en]
                  search_filter:
                    type: string
                    enum: [oem brand, product type]
                    description: >
                      the type of filtering criteria to further refine results.
                  search_filter_value:
                    oneOf:
                      - type: integer
                      - type: string
                    description: >
                      The value to be used in the search filter criteria depends on the
                      value used for the `search_filter`

                        | search_filter | value is a(n) | matching |
                        |---------------|------------|----------|
                        | oem brand id | integer | oem_brand_id of any matching product linked to the family
                        | product type | integer | product_type_id of any matching product linked to the family
                        
                  category_id:
                    type: array
                    description: >
                      The categories to add to the family search filtering. 
                      Specifying more than one category widens the search results. 
                      Note: The filtering actually looks for **the category on products linked to the family.**
                   
            examples:
              example1:
                description: >
                  Search for families having a `family_code` or `name_en` matching `F` (wildcard),

                  
                  **and** that are linked with products belonging to `category_id` = `5`, `6`, `7`, or `8`. 
                  
                  
                  **and** further apply a filter to get results where familes have any products associated with **product type** (`product_type_id`) = `2`.
                  
                  
                  Order the results by `id` (family id). 
                  

                  Do not skip any results, but limit the output to `10` entries.
                value:
                  offset: 0
                  limit: 10
                  order_by:
                    - id
                  search_term: F
                  search_term_fields:
                    - family_code
                    - name_en
                  category_id:
                    - 4
                    - 5
                    - 6
                    - 7
                  search_filter: product type
                  search_filter_value: 2
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    description: The total number of matching families in the database.
                    type: integer
                  product_views:
                    type: array
                    items:
                      $ref: '#/components/schemas/FamilyView'
            text/csv: {} 
        400:
          $ref: '#/components/responses/400ClientError'
        500:
          $ref: '#/components/responses/500ServerError'
  /families/search/download:
    post:
      summary: Download advanced search results for families.
      description: Extends basic search capabilities, allowing a search term to be used across different family fields. Additional filter criteria can be used to further narrow the search-term search.
      tags:
        - families
      requestBody:
        description: >
          The search criteria to use for the search.
        required: true
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/SearchCriteria'
              - type: object
                properties:
                  search_term:
                    type: string
                    description: wildcard search value to be used for searching any/all of the `search_term_fields`
                  search_term_fields:
                    description: which family fields the search term is applied to
                    type: array
                    items:
                      type: string
                      enum: [family_code, oem_brand_en, name_en, category_en]
                  search_filter:
                    type: string
                    enum: [oem brand, product type]
                    description: >
                      the type of filtering criteria to further refine results.
                  search_filter_value:
                    oneOf:
                      - type: integer
                      - type: string
                    description: >
                      The value to be used in the search filter criteria depends on the
                      value used for the `search_filter`

                        | search_filter | value is a(n) | matching |
                        |---------------|------------|----------|
                        | oem brand id | integer | oem_brand_id of any matching product linked to the family
                        | product type | integer | product_type_id of any matching product linked to the family
                        
                  category_id:
                    type: array
                    description: >
                      The categories to add to the family search filtering. 
                      Specifying more than one category widens the search results. 
                      Note: The filtering actually looks for **the category on products linked to the family.**
                   
            examples:
              example1:
                description: >
                  Search for families having a `family_code` or `name_en` matching `F` (wildcard),

                  
                  **and** that are linked with products belonging to `category_id` = `5`, `6`, `7`, or `8`. 
                  
                  
                  **and** further apply a filter to get results where familes have any products associated with **product type** (`product_type_id`) = `2`.
                  
                  
                  Order the results by `id` (family id). 
                  

                  Do not skip any results, but limit the output to `10` entries.
                value:
                  offset: 0
                  limit: 10
                  order_by:
                    - id
                  search_term: F
                  search_term_fields:
                    - family_code
                    - name_en
                  category_id:
                    - 4
                    - 5
                    - 6
                    - 7
                  search_filter: product type
                  search_filter_value: 2
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    description: The total number of matching families in the database.
                    type: integer
                  family_views:
                    type: array
                    items:
                      $ref: '#/components/schemas/FamilyView'
                  headers:
                    type: array
                    items:
                      type: object
                      description: a single object whose properties can be used to create headers for CSV/table rendering.
            text/csv: {} 
        400:
          $ref: '#/components/responses/400ClientError'
        500:
          $ref: '#/components/responses/500ServerError'
...
components:
  schemas:
    ApiError:
      type: object
      properties:
        message:
          type: string
          description: A human-readable message describing the error.
        error:
          type: string
          description: The error details. Typically error details may contain sensitive information not intended to be shown directly in the user interface.
    Auditable:
      type: object
      properties:
        created:
          type: string
          format: date-time
          readOnly: true
        updated:
          type: string
          format: date-time
          readOnly: true
        version:
          type: integer
          readOnly: true
...
    Family:
      type: object
      description: The product family .
      allOf:
      - type: object
        properties:
          id:
            type: integer
            readOnly: true
          name_en:
            description: Name (English) of the family.
            type: string
            maxLength: 255
          family_code:
            type: string
            maxLength: 255
          family_connector_code:
            type: string
            maxLength: 255
          group_id:
            description: References the group.
            type: integer
          name_edit_user_id:
            description: user id who last edited the name
            type: integer
          name_edit_timestamp:
            description: timestamp of last name edit
            type: string
            format: date-time
          image_link_connector_distal:
            description: URL to a connector distal image.
            type: string
            maxLength: 255
          video_link:
            description: URL to a video for the family.
            type: string
            maxLength: 255
      - $ref: '#/components/schemas/Auditable'
    FamilyView:
      type: object
      description: The view of a family including extended information.
      allOf:
      - type: object
        properties:
          id:
            type: integer
            readOnly: true
          name_en:
            description: Name (English) of the family.
            type: string
            maxLength: 255
          family_code:
            type: string
            maxLength: 255
          family_connector_code:
            type: string
            maxLength: 255
          group_id:
            description: References the group.
            type: integer
          group_code:
            type: string
            maxLength: 255
          image_link_connector_distal:
            description: URL to a connector distal image.
            type: string
            maxLength: 255
          video_link:
            description: URL to a video for the family.
            type: string
            maxLength: 255
      - $ref: '#/components/schemas/Auditable'
...
    ProductFamily:
      type: object
      description: The family reference for a product.
      allOf:
      - type: object
        properties:
          id:
            type: integer
          product_id:
            type: integer
          family_id:
            type: integer
      - $ref: '#/components/schemas/Auditable'    
...
    SearchCriteria:
      type: object
      allOf:
      - type: object
        properties:
          offset:
            type: integer
            description: for pagination. A zero-indexed value indicating where to begin the output with in the set of all possible search results.
          limit:
            type: integer
            description: for pagination. The number of records to return in a single query (up to 1000).
          order_by:
            type: array
            description: Describes the ordering of the search results. Provide the field names to order by. Prefix the field name with a `-` if you want ordering to be in descending order.
            items:
              type: string
          search_term:
            type: string
            description: the search term value. It will be treated as a wildcard search within the `search_term_fields` values.
          search_term_fields:
            type: array
            description: the list of fields you want the `search_term` to apply to. The criteria will be applied as a (OR) when conducting the search across more than one field.
            items:
              type: string        
...
  responses:
    400ClientError:
      description: client error
      content:
        application/json:
          schema: 
            $ref: '#/components/schemas/ApiError'
...
    500ServerError:
      description: server error
      content:
        application/json:
          schema: 
            $ref: '#/components/schemas/ApiError'